preclinica crear,editar,listar

+++ lib/src/models/preclinica_model.dart
@@ -9,7 +9,7 @@ class Preclinica {
   int preclinicaId;
   int pacienteId;
   int doctorId;
-  int peso;
+  double peso;
   double altura;
   int frecuenciaRespiratoria;
   int ritmoCardiaco;
